Created accuracy calculation logic

+++ app/src/main/java/edu/konan_univ/senn/database/models/Question.java
@@ -7,12 +7,12 @@ public class Question {
 
     public String sentence;
 
-    public Double answer;
+    public Float answer;
 
     public Question() {
     }
 
-    public Question(String sentence, Double answer) {
+    public Question(String sentence, Float answer) {
         this.sentence = sentence;
         this.answer = answer;
     }
